+++ core/src/main/scala/better/files/File.scala
@@ -107,7 +107,9 @@ class File private (val path: Path)(implicit val fileSystem: FileSystem = path.g
     * If file does not exist (or is a directory) no change is done and the current file is returned
     */
   def changeExtensionTo(extension: String): File =
-    if (isRegularFile) renameTo(s"$nameWithoutExtension$extension") else this
+    if (isRegularFile) renameTo(s"$nameWithoutExtension$extension")
+    else if (notExists) File(s"$nameWithoutExtension$extension")
+    else this
 
   def contentType: Option[String] =
     Option(Files.probeContentType(path))