+++ core/src/test/scala/better/files/FileSpec.scala
@@ -200,31 +200,30 @@ class FileSpec extends CommonSpec {
 //    ls_r("core"/"src"/"test") should have length 8
 //  }
 
-//  NOTE: Commented out because it's no longer needed. Expected to be vetted and removed by the project maintainer.
-//  it should "support names/extensions" in {
-//    assume(isCI)
-//    fa.extension shouldBe None
-//    fa.nameWithoutExtension shouldBe fa.name
-//    t1.extension shouldBe Some(".txt")
-//    t1.extension(includeDot = false) shouldBe Some("txt")
-//    t3.extension shouldBe Some(".txt")
-//    t3.extension(includeAll = true) shouldBe Some(".scala.txt")
-//    t3.extension(includeDot = false, includeAll = true) shouldBe Some("scala.txt")
-//    t1.name shouldBe "t1.txt"
-//    t1.nameWithoutExtension shouldBe "t1"
-//    t1.changeExtensionTo(".md").name shouldBe "t1.md"
-//    (t1 < "hello world").changeExtensionTo(".txt").name shouldBe "t1.txt"
-//    t1.contentType shouldBe Some("text/plain")
-//    ("src" / "test").toString should include ("better-files")
-//    (t1 == t1.toString) shouldBe false
-//    (t1.contentAsString == t1.toString) shouldBe false
-//    (t1 == t1.contentAsString) shouldBe false
-//    t1.root shouldEqual fa.root
-//    file"/tmp/foo.scala.html".extension shouldBe Some(".html")
-//    file"/tmp/foo.scala.html".nameWithoutExtension shouldBe "foo"
-//    file"/tmp/foo.scala.html".nameWithoutExtension(includeAll = false) shouldBe "foo.scala"
-//    root.name shouldBe ""
-//  }
+  it should "support names/extensions" in {
+    assert(File("zzz").changeExtensionTo("ddd").name === "zzz.ddd")
+    fa.extension shouldBe None
+    fa.nameWithoutExtension shouldBe fa.name
+    t1.extension shouldBe Some(".txt")
+    t1.extension(includeDot = false) shouldBe Some("txt")
+    t3.extension shouldBe Some(".txt")
+    t3.extension(includeAll = true) shouldBe Some(".scala.txt")
+    t3.extension(includeDot = false, includeAll = true) shouldBe Some("scala.txt")
+    t1.name shouldBe "t1.txt"
+    t1.nameWithoutExtension shouldBe "t1"
+    t1.changeExtensionTo(".md").name shouldBe "t1.md"
+    (t1 < "hello world").changeExtensionTo(".txt").name shouldBe "t1.txt"
+    //t1.contentType shouldBe Some("text/plain")
+    ("src" / "test").toString should include("better-files")
+    (t1 == t1.toString) shouldBe false
+    (t1.contentAsString == t1.toString) shouldBe false
+    (t1 == t1.contentAsString) shouldBe false
+    t1.root shouldEqual fa.root
+    file"/tmp/foo.scala.html".extension shouldBe Some(".html")
+    file"/tmp/foo.scala.html".nameWithoutExtension shouldBe "foo"
+    file"/tmp/foo.scala.html".nameWithoutExtension(includeAll = false) shouldBe "foo.scala"
+    root.name shouldBe ""
+  }
 
   it should "hide/unhide" in {
     t1.isHidden shouldBe false